[DEVOPS-XXX] Convert release drafter workflow to be reusable (#186)
<!-- Please make sure you read the contribution guidelines and then fill
out the blanks below.

Please format the PR title appropriately based on the type of change:
  [JIRA-XXX]: <description>
-->

## Description

We've been having issues with the latest release of the release drafter
where it has been duplicating draft releases. This PR is to update the
release drafter workflow in this repository to convert it into a
reusable workflow so that we can link to it in other repositories and
fix the version.

## Related Links

<!-- List any links related to this pull request here

Replace "JIRA-XXX" with the your Jira issue key -->

- Jira Issue: N/A